requirements.txt
aiogram==3.*
python-dotenv
requests
loguru
pymysql
beautifulsoup4
lxml
